+++ nodejs/examples/.gitignore
@@ -1,1 +1,2 @@
 data/
+*.js
