+++ nodejs/examples/package.json
@@ -1,28 +1,38 @@
 {
-  "name": "examples",
-  "version": "1.0.0",
-  "description": "Examples for LanceDB",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Lance Devs",
-  "license": "Apache-2.0",
-  "dependencies": {
-    "@lancedb/lancedb": "file:../",
-    "@xenova/transformers": "^2.17.2"
-  },
-  "devDependencies": {
-    "typescript": "^5.5.4"
-  },
-  "compilerOptions": {
-    "target": "ESNext",
-    "module": "ESNext",
-    "moduleResolution": "Node",
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true
+    "name": "examples",
+    "version": "1.0.0",
+    "description": "Examples for LanceDB",
+    "main": "index.js",
+    "type": "module",
+    "scripts": {
+      "test": "jest",
+      "build": "tsc",
+      "lint": "biome check *.ts && biome format *.ts",
+      "lint-fix": "biome check --write *.ts && npm run format",
+      "format": "biome format --write *.ts"
+    },
+    "author": "Lance Devs",
+    "license": "Apache-2.0",
+    "dependencies": {
+      "@lancedb/lancedb": "file:../dist",
+      "@xenova/transformers": "^2.17.2",
+      "openai": "^4.29.2"
+    },
+    "devDependencies": {
+      "@biomejs/biome": "^1.7.3",
+      "@jest/globals": "^29.7.0",
+      "jest": "^29.7.0",
+      "ts-jest": "^29.2.5",
+      "typescript": "^5.5.4"
+    },
+    "compilerOptions": {
+      "target": "ESNext",
+      "module": "ESNext",
+      "moduleResolution": "Node",
+      "strict": true,
+      "esModuleInterop": true,
+      "skipLibCheck": true,
+      "forceConsistentCasingInFileNames": true
+    }
   }
-}
+
