+++ nodejs/examples/tsconfig.json
@@ -1,0 +1,18 @@
+{
+    "include": ["*.test.ts"],
+    "compilerOptions": {
+      "target": "es2022",
+      "module": "NodeNext",
+      "declaration": true,
+      "outDir": "./dist",
+      "strict": true,
+      "allowJs": true,
+      "resolveJsonModule": true,
+      "emitDecoratorMetadata": true,
+      "experimentalDecorators": true,
+      "moduleResolution": "NodeNext",
+      "allowImportingTsExtensions": true,
+      "emitDeclarationOnly": true
+    }
+  }
+
